Mosa.DeviceDriver: structural refactoring
