+++ Source/Mosa.DeviceDriver/PCI/VirtIO/VirtIOFlags.cs
@@ -1,6 +1,6 @@
 // Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-namespace Mosa.DeviceSystem.VirtIO;
+namespace Mosa.DeviceDriver.PCI.VirtIO;
 
 /// <summary>
 /// Generic VirtIO flags to send to a device for initialization.
