+++ Source/Mosa.DeviceDriver/PCI/VirtIO/VirtIOPCICommonConfiguration.cs
@@ -1,6 +1,6 @@
 // Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-namespace Mosa.DeviceSystem.VirtIO;
+namespace Mosa.DeviceDriver.PCI.VirtIO;
 
 /// <summary>
 /// Generic VirtIO flags to send to a device, most notably for querying data and dealing with virtqueues.
