+++ Source/Mosa.DeviceDriver/PCI/VirtIO/VirtIOQueueDescriptorFlags.cs
@@ -1,6 +1,6 @@
 // Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-namespace Mosa.DeviceSystem.VirtIO;
+namespace Mosa.DeviceDriver.PCI.VirtIO;
 
 /// <summary>
 /// Flags used in a VirtIO descriptor. "HasNext" tells the driver that a next descriptor is available (essentially making it a chain),
